New improved CausalTestingFramework class
