+++ causal_testing/estimation/cubic_spline_estimator.py
@@ -35,6 +35,7 @@ class CubicSplineRegressionEstimator(LinearRegressionEstimator):
         formula: str = None,
         alpha: float = 0.05,
         expected_relationship=None,
+        adjustment_config: dict[Variable, Any] = None,
     ):
         super().__init__(
             base_test_case=base_test_case,
@@ -47,9 +48,10 @@ class CubicSplineRegressionEstimator(LinearRegressionEstimator):
         )
 
         self.expected_relationship = expected_relationship
+        self.adjustment_config = adjustment_config
 
         if effect_modifiers is None:
-            effect_modifiers = []
+            effect_modifiers = {}
 
         if formula is None:
             terms = (
@@ -66,24 +68,24 @@ class CubicSplineRegressionEstimator(LinearRegressionEstimator):
         model = self.regressor(formula=self.formula, data=df).fit(disp=0)
         return model
 
-    def estimate_ate_calculated(self, df: pd.DataFrame, adjustment_config: dict = None) -> EffectEstimate:
+    def estimate_ate_calculated(
+        self,
+        df: pd.DataFrame,
+    ) -> EffectEstimate:
         """Estimate the ate effect of the treatment on the outcome. That is, the change in outcome caused
         by changing the treatment variable from the control value to the treatment value. Here, we actually
         calculate the expected outcomes under control and treatment and divide one by the other. This
         allows for custom terms to be put in such as squares, inverses, products, etc.
 
         :param df: The data to use.
-        :param adjustment_config: The configuration of the adjustment set as a dict mapping variable names to
-                                   their values. N.B. Every variable in the adjustment set MUST have a value in
-                                   order to estimate the outcome under control and treatment.
 
         :return: The average treatment effect.
         """
         model = self.fit_model(df)
 
         x = pd.DataFrame({"Intercept": [1], self.base_test_case.treatment_variable.name: [self.treatment_value]})
-        if adjustment_config is not None:
-            for k, v in adjustment_config.items():
+        if self.adjustment_config is not None:
+            for k, v in self.adjustment_config.items():
                 x[k] = v
         if self.effect_modifiers is not None:
             for k, v in self.effect_modifiers.items():
