+++ tests/main_tests/test_main.py
@@ -7,128 +7,50 @@ import shutil
 import json
 import pandas as pd
 
-from causal_testing.main import CausalTestingPaths, CausalTestingFramework
+from causal_testing.causal_testing_framework import CausalTestingFramework
 from causal_testing.__main__ import main
-
-
-class TestCausalTestingPaths(unittest.TestCase):
-
-    def setUp(self):
-        self.dag_path = "tests/resources/data/dag.dot"
-        self.data_paths = ["tests/resources/data/data.csv"]
-        self.test_config_path = "tests/resources/data/tests.json"
-        self.output_path = Path("results/results.json")
-        self.include_edges_path = Path("tests/resources/data/include_edges.dot")
-        self.exclude_edges_path = Path("tests/resources/data/exclude_edges.dot")
-
-    def test_missing_dag(self):
-        with self.assertRaises(FileNotFoundError) as e:
-            CausalTestingPaths("missing.dot", self.data_paths, self.test_config_path, self.output_path).validate_paths()
-        self.assertEqual("DAG file not found: missing.dot", str(e.exception))
-
-    def test_missing_data(self):
-        with self.assertRaises(FileNotFoundError) as e:
-            CausalTestingPaths(self.dag_path, ["missing.csv"], self.test_config_path, self.output_path).validate_paths()
-        self.assertEqual("Data file not found: missing.csv", str(e.exception))
-
-    def test_missing_tests(self):
-        with self.assertRaises(FileNotFoundError) as e:
-            CausalTestingPaths(self.dag_path, self.data_paths, "missing.json", self.output_path).validate_paths()
-        self.assertEqual("Test configuration file not found: missing.json", str(e.exception))
-
-    def test_output_file_created(self):
-        self.assertFalse(self.output_path.parent.exists())
-        CausalTestingPaths(self.dag_path, self.data_paths, self.test_config_path, self.output_path).validate_paths()
-        self.assertTrue(self.output_path.parent.exists())
-
-    def test_missing_include_edges(self):
-        with self.assertRaises(FileNotFoundError) as e:
-            CausalTestingPaths(
-                self.dag_path,
-                self.data_paths,
-                self.test_config_path,
-                self.output_path,
-                "missing.dot",
-                self.exclude_edges_path,
-            ).validate_paths()
-        self.assertEqual("Data file not found: missing.dot", str(e.exception))
-
-    def test_missing_exclude_edges(self):
-        with self.assertRaises(FileNotFoundError) as e:
-            CausalTestingPaths(
-                self.dag_path,
-                self.data_paths,
-                self.test_config_path,
-                self.output_path,
-                self.include_edges_path,
-                "missing.dot",
-            ).validate_paths()
-        self.assertEqual("Data file not found: missing.dot", str(e.exception))
-
-    def tearDown(self):
-        if self.output_path.parent.exists():
-            shutil.rmtree(self.output_path.parent)
 
 
 class TestCausalTestingFramework(unittest.TestCase):
     def setUp(self):
         self.dag_path = "tests/resources/data/dag.dot"
         self.data_paths = ["tests/resources/data/data.csv"]
-        self.test_config_path = "tests/resources/data/tests.json"
+        self.test_cases_path = "tests/resources/data/tests.json"
         self.output_path = Path("results/results.json")
         self.include_edges_path = "tests/resources/data/include_edges.dot"
         self.exclude_edges_path = "tests/resources/data/exclude_edges.dot"
-        self.paths = CausalTestingPaths(
-            dag_path=self.dag_path,
-            data_paths=self.data_paths,
-            test_config_path=self.test_config_path,
-            output_path=self.output_path,
-            include_edges_path=self.include_edges_path,
-            exclude_edges_path=self.exclude_edges_path,
-        )
+        self.paths = {
+            "dag_path": self.dag_path,
+            "data_paths": self.data_paths,
+            "test_cases_path": self.test_cases_path,
+        }
 
     def test_load_data(self):
-        csv_framework = CausalTestingFramework(self.paths)
-        csv_df = csv_framework.load_data()
+        csv_framework = CausalTestingFramework()
+        csv_framework.load_data(self.data_paths)
 
-        pqt_framework = CausalTestingFramework(
-            CausalTestingPaths(
-                dag_path=self.dag_path,
-                data_paths=["tests/resources/data/data.pqt"],
-                test_config_path=self.test_config_path,
-                output_path=self.output_path,
-            )
-        )
-        pqt_df = pqt_framework.load_data()
-        pd.testing.assert_frame_equal(csv_df, pqt_df)
-
-    def test_load_data_invalid(self):
-        framework = CausalTestingFramework(
-            CausalTestingPaths(
-                dag_path=self.dag_path,
-                data_paths=[self.dag_path],
-                test_config_path=self.test_config_path,
-                output_path=self.output_path,
-            )
-        )
-        with self.assertRaises(ValueError):
-            framework.load_data()
+        pqt_framework = CausalTestingFramework()
+        pqt_framework.load_data([path.replace(".csv", ".pqt") for path in self.data_paths])
+        pd.testing.assert_frame_equal(csv_framework.df, pqt_framework.df)
 
     def test_load_data_query(self):
-        framework = CausalTestingFramework(self.paths)
-        self.assertFalse((framework.load_data()["test_input"] > 4).all())
-        self.assertTrue((framework.load_data("test_input > 4")["test_input"] > 4).all())
+        framework = CausalTestingFramework()
+        framework.load_data(data_paths=self.data_paths)
+        self.assertFalse((framework.df["test_input"] > 4).all())
+
+        framework.load_data(data_paths=self.data_paths, query="test_input > 4")
+        self.assertTrue((framework.df["test_input"] > 4).all())
 
     def test_load_dag_missing_node(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         framework.dag.add_node("missing")
         with self.assertRaises(ValueError):
             framework.create_variables()
 
     def test_create_base_test_case_missing_treatment(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         with self.assertRaises(KeyError) as e:
             framework.create_base_test(
                 {"treatment_variable": "missing", "expected_effect": {"test_outcome": "NoEffect"}}
@@ -136,17 +58,25 @@ class TestCausalTestingFramework(unittest.TestCase):
         self.assertEqual("\"Treatment variable 'missing' not found in inputs or outputs\"", str(e.exception))
 
     def test_create_base_test_case_missing_estimator(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         with self.assertRaises(ValueError) as e:
-            framework.create_causal_test({}, None)
+            framework.create_causal_test(
+                {"treatment_variable": "test_input", "expected_effect": {"test_output": "NoEffect"}}
+            )
         self.assertEqual("Test configuration must specify an estimator", str(e.exception))
 
     def test_create_test_case_invalid_estimator(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         with self.assertRaises(ValueError) as e:
-            framework.create_causal_test({"estimator": "InvalidEstimator"}, None)
+            framework.create_causal_test(
+                {
+                    "treatment_variable": "test_input",
+                    "expected_effect": {"test_output": "NoEffect"},
+                    "estimator": "InvalidEstimator",
+                }
+            )
         self.assertEqual(
             f"Unsupported estimator InvalidEstimator. Supported: ['CubicSplineEstimator', 'IPCWEstimator', 'InstrumentalVariableEstimator', 'LinearRegressionEstimator', 'LogisticRegressionEstimator', 'MultinomialRegressionEstimator']. "
             "If you have implemented a custom estimator, you will need to add this to your entrypoints via your "
@@ -155,8 +85,8 @@ class TestCausalTestingFramework(unittest.TestCase):
         )
 
     def test_create_test_case_invalid_effect(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         test = {
             "name": "test1",
             "treatment_variable": "test_input",
@@ -166,7 +96,7 @@ class TestCausalTestingFramework(unittest.TestCase):
         }
         base_test_case = framework.create_base_test(test)
         with self.assertRaises(ValueError) as e:
-            framework.create_causal_test(test, base_test_case)
+            framework.create_causal_test(test)
         self.assertEqual(
             f"Unsupported causal effect InvalidEffect. Supported: ['ExactValue', 'Negative', 'NoEffect', 'Positive', 'SomeEffect']. "
             "If you have implemented a custom causal effect, you will need to add this to your entrypoints via your "
@@ -175,8 +105,8 @@ class TestCausalTestingFramework(unittest.TestCase):
         )
 
     def test_create_test_case_effect_kwargs(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         test = {
             "name": "test1",
             "treatment_variable": "test_input",
@@ -186,12 +116,12 @@ class TestCausalTestingFramework(unittest.TestCase):
             "effect_kwargs": {"value": 4},
         }
         base_test_case = framework.create_base_test(test)
-        test_case = framework.create_causal_test(test, base_test_case)
+        test_case = framework.create_causal_test(test)
         self.assertEqual(test_case.expected_causal_effect.value, 4)
 
     def test_create_test_case_estimator_kwargs(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         test = {
             "name": "test1",
             "treatment_variable": "test_input",
@@ -201,71 +131,67 @@ class TestCausalTestingFramework(unittest.TestCase):
             "estimator_kwargs": {"instrument": "instrumental_variable"},
         }
         base_test_case = framework.create_base_test(test)
-        test_case = framework.create_causal_test(test, base_test_case)
+        test_case = framework.create_causal_test(test)
         self.assertEqual(test_case.estimator.instrument, "instrumental_variable")
 
     def test_create_base_test_case_missing_outcome(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         with self.assertRaises(KeyError) as e:
             framework.create_base_test({"treatment_variable": "test_input", "expected_effect": {"missing": "NoEffect"}})
         self.assertEqual("\"Outcome variable 'missing' not found in inputs or outputs\"", str(e.exception))
 
     def test_unloaded_tests(self):
-        framework = CausalTestingFramework(self.paths)
+        framework = CausalTestingFramework()
         with self.assertRaises(ValueError) as e:
             framework.run_tests()
-        self.assertEqual("No tests loaded. Call load_tests() first.", str(e.exception))
+        self.assertEqual("No tests to run.", str(e.exception))
 
     def test_ctf(self):
-        framework = CausalTestingFramework(self.paths)
-        framework.setup()
-
-        framework.load_tests()
+        framework = CausalTestingFramework()
+        framework.setup(**self.paths)
         framework.run_tests()
-        json_results = framework.save_results()
+        json_results = framework.save_results(self.output_path)
 
-        with open(self.test_config_path, "r", encoding="utf-8") as f:
+        with open(self.test_cases_path, "r", encoding="utf-8") as f:
             test_configs = json.load(f)
 
-            self.assertEqual(len(json_results), len(test_configs["tests"]))
+        self.assertEqual(len(json_results), len(test_configs["tests"]))
 
-            result_index = 0
-            for i, test_config in enumerate(test_configs["tests"]):
-                result = json_results[i]
+        result_index = 0
+        for i, test_config in enumerate(test_configs["tests"]):
+            result = json_results[i]
 
-                if test_config.get("skip", False):
-                    self.assertEqual(result["skip"], True)
-                    self.assertEqual(result["passed"], None)
-                    self.assertEqual(result["result"]["status"], "skipped")
-                else:
-                    test_case = framework.test_cases[result_index]
-                    result_index += 1
+            if test_config.get("skip", False):
+                self.assertEqual(result["skip"], True)
+                self.assertEqual(result["passed"], None)
+                self.assertEqual(result["result"]["status"], "skipped")
+            else:
+                test_case = framework.test_cases[result_index]
+                result_index += 1
 
-                    test_passed = (
-                        test_case.expected_causal_effect.apply(test_case.result)
-                        if test_case.result.effect_estimate is not None
-                        else False
-                    )
-                    self.assertEqual(result["passed"], test_passed)
+                test_passed = (
+                    test_case.expected_causal_effect.apply(test_case.result)
+                    if test_case.result.effect_estimate is not None
+                    else False
+                )
+                self.assertEqual(result["passed"], test_passed)
 
     def test_ctf_exception(self):
-        framework = CausalTestingFramework(self.paths, query="test_input < 0")
-        framework.setup()
+        framework = CausalTestingFramework(self.paths)
+        framework.setup(**self.paths, query="test_input < 0")
 
-        framework.load_tests()
         with self.assertRaises(ValueError):
             framework.run_tests()
 
     def test_ctf_exception_silent(self):
-        framework = CausalTestingFramework(self.paths, query="test_input < 0")
-        framework.setup()
+        framework = CausalTestingFramework(self.paths)
+        framework.setup(**self.paths, query="test_input < 0")
 
-        framework.load_tests()
         framework.run_tests(silent=True)
-        json_results = framework.save_results()
+        json_results = framework.save_results(self.output_path)
 
-        with open(self.test_config_path, "r", encoding="utf-8") as f:
+        with open(self.test_cases_path, "r", encoding="utf-8") as f:
             test_configs = json.load(f)
 
             non_skipped_configs = [t for t in test_configs["tests"] if not t.get("skip", False)]
@@ -287,7 +213,7 @@ class TestCausalTestingFramework(unittest.TestCase):
                 "--data-paths",
                 str(self.data_paths[0]),
                 "--test-config",
-                str(self.test_config_path),
+                str(self.test_cases_path),
                 "--output",
                 str(self.output_path.parent / "main.json"),
             ],
@@ -306,7 +232,7 @@ class TestCausalTestingFramework(unittest.TestCase):
                 "--data-paths",
                 str(self.data_paths[0]),
                 "--test-config",
-                str(self.test_config_path),
+                str(self.test_cases_path),
                 "--output",
                 str(self.output_path.parent / "main.json"),
                 "-a",
@@ -329,7 +255,7 @@ class TestCausalTestingFramework(unittest.TestCase):
                 "--data-paths",
                 str(self.data_paths[0]),
                 "--test-config",
-                str(self.test_config_path),
+                str(self.test_cases_path),
                 "--output",
                 str(self.output_path.parent / "main.json"),
                 "-b",
@@ -353,7 +279,7 @@ class TestCausalTestingFramework(unittest.TestCase):
                 "--data-paths",
                 str(self.data_paths[0]),
                 "--test-config",
-                str(self.test_config_path),
+                str(self.test_cases_path),
                 "--output",
                 str(self.output_path.parent / "main.json"),
                 "-a",
